# contracts/openapi/api.json
{
  "components": {
    "schemas": {
...
      "Body_knowledge_upload_document": {
        "properties": {
          "upload": {
            "format": "binary",
            "title": "Upload",
            "type": "string"
          }
        },
        "required": [
          "upload"
        ],
        "title": "Body_knowledge_upload_document",
        "type": "object"
      },
...
      "DocumentAccepted": {
        "additionalProperties": false,
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentSummary"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "jobId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Jobid",
            "type": "string"
          }
        },
        "required": [
          "document",
          "jobId",
          "duplicate"
        ],
        "title": "DocumentAccepted",
        "type": "object"
      },
...
      "DocumentDetail": {
        "additionalProperties": false,
        "properties": {
          "chunkCount": {
            "maximum": 10000,
            "minimum": 0,
            "title": "Chunkcount",
            "type": "integer"
          },
          "documentId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Documentid",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errorcode"
          },
          "errorSummary": {
            "anyOf": [
              {
                "maxLength": 240,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errorsummary"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "mediaType": {
            "enum": [
              "application/pdf",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "text/markdown",
              "text/plain"
            ],
            "title": "Mediatype",
            "type": "string"
          },
          "normalizedPreview": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Normalizedpreview"
          },
          "revisionId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Revisionid",
            "type": "string"
          },
          "sourceContentHash": {
            "maxLength": 71,
            "minLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Sourcecontenthash",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/IngestionStage"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/DocumentStageSnapshot"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updatedAt": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "filename",
          "mediaType",
          "status",
          "stage",
          "chunkCount",
          "updatedAt",
          "revisionId",
          "sourceContentHash",
          "stages"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "maxItems": 50,
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DocumentPage",
        "type": "object"
      },
...
      "DocumentStatus": {
        "enum": [
          "queued",
          "processing",
          "ready",
          "failed",
          "deleting"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentSummary": {
        "additionalProperties": false,
        "properties": {
          "chunkCount": {
            "maximum": 10000,
            "minimum": 0,
            "title": "Chunkcount",
            "type": "integer"
          },
          "documentId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Documentid",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errorcode"
          },
          "errorSummary": {
            "anyOf": [
              {
                "maxLength": 240,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errorsummary"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "mediaType": {
            "enum": [
              "application/pdf",
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
              "text/markdown",
              "text/plain"
            ],
            "title": "Mediatype",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/IngestionStage"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updatedAt": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "documentId",
          "filename",
          "mediaType",
          "status",
          "stage",
          "chunkCount",
          "updatedAt"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
...
      "IngestionStage": {
        "enum": [
          "stored",
          "parsing",
          "chunking",
          "embedding",
          "publishing",
          "ready"
        ],
        "title": "IngestionStage",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "TAP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/knowledge/documents": {
      "get": {
        "operationId": "knowledge_list_documents",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid list request"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "List Documents",
        "tags": [
          "knowledge"
        ]
      },
      "post": {
        "operationId": "knowledge_upload_document",
        "parameters": [
          {
            "in": "header",
            "name": "content-length",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Length"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_knowledge_upload_document"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid document upload"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{document_id}": {
      "delete": {
        "operationId": "knowledge_delete_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid document ID"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "knowledge"
        ]
      },
      "get": {
        "operationId": "knowledge_get_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid document ID"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Get Document",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/knowledge/documents/{document_id}/retry": {
      "post": {
        "operationId": "knowledge_retry_document",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid document ID"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Retry Document",
        "tags": [
          "knowledge"
        ]
      }
    }
  }
}